+++ src/scicloj/noj/v1/datasets.clj
@@ -1,0 +1,10 @@
+(ns scicloj.noj.v1.datasets
+  (:require [tablecloth.api :as tc]))
+
+(def mtcars
+  (-> "data/mtcars.csv"
+      (tc/dataset {:key-fn keyword})))
+
+(def iris
+  (-> "data/iris.csv"
+      (tc/dataset {:key-fn keyword})))